+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+#Changelog
+
+## 0.2.3 (2016-03-21)
+
+Bugfixes:
+
+    - remove `href` attribute from DefaultSorter
+    - add style `cursor: pointer` to links in DefaultSorter and BootstrapPaginator
+    
 ## 0.2.2 (2016-03-21)
 
 Bugfixes:
 
     - remove `href` attribute from BootstrapPaginator template
+